feat: pinning gmail message link for fast-open
